#79 Multi-line and comment support some fixes
