+++ jxls/src/main/java/org/jxls/builder/xls/XlsCommentAreaBuilder.java
@@ -26,6 +26,7 @@ import org.jxls.command.UpdateCellCommand;
 import org.jxls.common.AreaRef;
 import org.jxls.common.CellData;
 import org.jxls.common.CellRef;
+import org.jxls.extractors.LiteralsExtractor;
 import org.jxls.transform.Transformer;
 import org.jxls.util.Util;
 import org.slf4j.Logger;
@@ -94,7 +95,7 @@ public class XlsCommentAreaBuilder implements AreaBuilder {
     public static final String COMMAND_PREFIX = "jx:";
     private static final String ATTR_PREFIX = "(";
     private static final String ATTR_SUFFIX = ")";
-    public static final String LINE_SEPARATOR = "__LINE_SEPARATOR__";
+    public static final String LINE_SEPARATOR = "__LINE_SEPARATOR__";    
     /*
      * In addition to normal (straight) single and double quotes, this regex
      * includes the following commonly occurring quote-like characters (some
@@ -223,97 +224,12 @@ public class XlsCommentAreaBuilder implements AreaBuilder {
         }
         return userAreas;
     }
-
-    private void extractLiterals(CellData cellData, String text, List<String> literalList) {
-		String literal = "";
-		for (int i = 0, n = text.length(); i < n; i++) {
-
-			// if the first 3 chars of literal are 'jx:'
-			if (literal.length() >= 3
-					&& literal.substring(literal.length() - 3, literal.length()).equals(COMMAND_PREFIX)) {
-
-				// save comment parsed before
-				literal = literal.substring(0, literal.length() - 3);
-				makeCommentLines(literal, literalList);
-				String remainingText = text.substring(i - 3, n);
-				String cmd = makeOperators(cellData, remainingText);
-				literal = "";
-				literalList.add(cmd);
-				i = i + cmd.length() - 3;
-
-				// if text no ending
-				if (i != n)
-					literal = String.valueOf(text.charAt(i));
-				else
-					literal = "";
-
-			} else {
-				// else is the next char to new literal
-				literal = literal + String.valueOf(text.charAt(i));
-			}
-		}
-
-		// last comment
-		if (!literal.isEmpty()) {
-			makeCommentLines(literal, literalList);
-		}
-	}
     
-	private void makeCommentLines(String literal, List<String> literalList) {
-		literalList.addAll(
-				Arrays.stream(literal.split("\n"))
-					.filter(line -> ! line.isEmpty())
-					.collect(toList()));
-	}
-	
-	private String makeOperators(CellData cellData, String text) {
-		// result command
-		String cmd = "";
-		Stack<Character> bracketsStack = new Stack<>();
-		Stack<Character> quotationsStack = new Stack<>();
-
-		for (int i = 0; i < text.length(); i++) {
-			char symbol = text.charAt(i);
-
-			if (symbol == '"' || symbol == '\'')
-				if (!quotationsStack.empty() && quotationsStack.peek() == symbol)
-					quotationsStack.pop();
-				else
-					quotationsStack.push(symbol);
-
-			switch (symbol) {
-			case '(':
-				if (quotationsStack.empty())
-					bracketsStack.push(symbol);
-				break;
-			case ')':
-				if (quotationsStack.empty()) {
-					if (bracketsStack.peek() == '(') {
-						bracketsStack.pop();
-						if (bracketsStack.empty())
-							return cmd + symbol;
-					} else {
-						// no opening, but closing !!
-						String errorMessage = "Got closing ) but no opening of it. Pos: " + text.substring(0, i);
-						logger.error(errorMessage);
-						throw new IllegalStateException(errorMessage);
-					}
-				}
-				break;
-			default:
-				break;
-			}
-
-			cmd = cmd + symbol;
-
-		}
-
-		return cmd;
-	}
     
     private List<CommandData> buildCommands(CellData cellData, String text) {
         List<String> literalList = new ArrayList<String>();
-        extractLiterals(cellData, text, literalList);
+		LiteralsExtractor extractor = new LiteralsExtractor(text, literalList);
+    	literalList = extractor.extract();
 
         List<CommandData> commandDatas = new ArrayList<CommandData>();
         for (String commentLine : literalList) {
@@ -326,7 +242,6 @@ public class XlsCommentAreaBuilder implements AreaBuilder {
                 int nameEndIndex = line.indexOf(ATTR_PREFIX, COMMAND_PREFIX.length());
                 if (nameEndIndex < 0) {
                     String errMsg = "Failed to parse command line [" + line + "]. Expected '" + ATTR_PREFIX + "' symbol.";
-                    logger.error(errMsg);
                     throw new IllegalStateException(errMsg);
                 }
                 String commandName = line.substring(COMMAND_PREFIX.length(), nameEndIndex).trim();
